fix(frontend): remove unnecessary uploadButton component tm-42
