--- apps/frontend/src/libs/components/upload-button/styles.module.css
@@ -1,39 +1,0 @@
-.basic {
-	color: var(--main-white);
-	background-color: transparent;
-	border-color: var(--main-white);
-}
-
-.basic:hover {
-	color: var(--color-text-300);
-	cursor: pointer;
-	background-color: var(--main-white);
-}
-
-.small {
-	column-gap: 4px;
-	min-height: 32px;
-	padding: 4px 16px;
-	font-size: var(--font-size-200);
-	border-radius: 26px;
-}
-
-.noImage {
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	width: 100%;
-	height: 100%;
-	padding: 28px;
-	font-size: 12px;
-	color: var(--main-white);
-	text-align: center;
-	background-color: var(--color-menu);
-	border: 2px dashed var(--main-white);
-	border-radius: 50%;
-}
-
-.noImage:hover {
-	cursor: pointer;
-	background-color: var(--color-input);
-}
