--- apps/frontend/src/pages/profile/components/avatar/styles.module.css
@@ -1,13 +1,0 @@
-.avatar {
-	width: 133px;
-	height: 133px;
-	overflow: hidden;
-	border-radius: 50%;
-}
-
-.avatarImage {
-	width: 100%;
-	height: 100%;
-	object-fit: cover;
-	user-select: none;
-}
