--- apps/frontend/src/pages/profile/components/components.ts
@@ -1,2 +1,0 @@
-export { Avatar } from "./avatar/avatar.js";
-export { UploadAvatarButton } from "./upload-avatar-button/upload-avatar-button.js";
