+++ apps/frontend/src/pages/profile/profile.tsx
@@ -1,5 +1,6 @@
 import profileCharacter from "~/assets/img/user-details-img.png";
 import { Button, Image, Input } from "~/libs/components/components.js";
+import { DEFAULT_USER_AVATAR } from "~/libs/constants/constants.js";
 import { AppRoute } from "~/libs/enums/enums.js";
 import {
 	useAppDispatch,
@@ -7,7 +8,7 @@ import {
 	useAppSelector,
 	useCallback,
 	useParams,
-	useState,
+	useRef,
 } from "~/libs/hooks/hooks.js";
 import {
 	type UserAuthResponseDto,
@@ -16,15 +17,14 @@ import {
 	actions as usersActions,
 } from "~/modules/users/users.js";
 
-import { Avatar, UploadAvatarButton } from "./components/components.js";
 import styles from "./styles.module.css";
 
 const Profile: React.FC = () => {
-	const [selectedFile, setSelectedFile] = useState<File | null>(null);
 	const { userId } = useParams<{ userId: string }>();
 	const user = useAppSelector(({ auth }) => {
 		return auth.user as UserAuthResponseDto;
 	});
+	const inputReference = useRef<HTMLInputElement>(null);
 	const dispatch = useAppDispatch();
 
 	const defaultValues = {
@@ -37,10 +37,13 @@ const Profile: React.FC = () => {
 		validationSchema: userProfileValidationSchema,
 	});
 
+	const handleClick = useCallback((): void => {
+		inputReference.current?.click();
+	}, []);
+
 	const handleFileChange = useCallback(
 		(event: React.ChangeEvent<HTMLInputElement>): void => {
 			const [file] = event.target.files ?? [];
-			setSelectedFile(file ?? null);
 
 			if (file) {
 				const formData = new FormData();
@@ -80,15 +83,29 @@ const Profile: React.FC = () => {
 				<span className={styles["profile-title"]}>My profile</span>
 				<form name="profile" onSubmit={handleFormSubmit}>
 					<div className={styles["avatar-container"]}>
-						<Avatar
-							onFileChange={handleFileChange}
-							selectedFile={selectedFile}
-						/>
-						<div>
-							<UploadAvatarButton onFileChange={handleFileChange} />
+						<div className={styles["profile-image-wrapper"]}>
+							<Image
+								alt="avatar"
+								className={styles["profile-image"]}
+								src={user.avatarUrl ?? DEFAULT_USER_AVATAR}
+							/>
 						</div>
+						<button
+							className={styles["upload-button"]}
+							onClick={handleClick}
+							type="button"
+						>
+							<span>Change photo</span>
+							<input
+								accept={"image/*"}
+								hidden
+								onChange={handleFileChange}
+								ref={inputReference}
+								type="file"
+							/>
+						</button>
 					</div>
-					<div className={styles["profile-form"]}>
+					<div className={styles["profile-inputs"]}>
 						<fieldset className={styles["fieldset"]}>
 							<Input
 								color="light"
